Rename EditButton file and remove unused signup files
This commit renames the EditButton.jsx file to EditProfileButton.jsx. It also deletes all files within the old signup modal.
